Create First script
TASK : Write a script that will take a single filename as an argument and add execute permission (+x) to the file for the user, but only if the file is a regular file. 
Your script must check to see that there is exactly one argument. 
If there are no arguments or more than one argument, your script must produce a "usage" message that tells how to use the script.
Your shell script must check to see if the file exists; if it doesn’t, give an error message. 
You must check to see if the file is a regular file; if not, give an appropriate message.