Create Second script
TASK : Write a script that will do an ls -l command on each of its arguments only if the argument is the name of a regular file. 
This script will have a for loop, and inside the loop will be an if-test that uses one of the file test options.
Your program must work on any list of files from the command line.
Here’s how it will work: presume you have files named xyz and abc, and a directory named mydir. If you run the script:
./listfiles.sh xyz mydir abc nosuchfile
You will get output like this, since xyz and abc are the only two regular files:

-rw-r--r-- 1 linux199 users 24490 2006-11-21 15:26 xyz
-rw-r-xrw- 1 linux199 users 5963 2006-11-16 18:47 abc